refactor proposal workflow evidence routes
