+++ src/app/router_registry.py
@@ -128,6 +128,9 @@ from app.routers.proposal_narratives import router as proposal_narratives_router
 from app.routers.proposal_operations import router as proposal_operations_router
 from app.routers.proposal_versions import router as proposal_versions_router
 from app.routers.proposal_workflow import router as proposal_workflow_router
+from app.routers.proposal_workflow_evidence import (
+    router as proposal_workflow_evidence_router,
+)
 from app.routers.proposals import router as proposals_router
 from app.routers.reporting import router as reporting_router
 from app.routers.reporting_batch_controls import (
@@ -186,6 +189,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(proposal_operations_router)
     app.include_router(proposal_versions_router)
     app.include_router(proposal_workflow_router)
+    app.include_router(proposal_workflow_evidence_router)
     app.include_router(proposal_narratives_router)
     app.include_router(proposal_delivery_router)
     app.include_router(proposal_memos_router)
