+++ src/app/routers/proposal_workflow.py
@@ -2,11 +2,8 @@ from fastapi import APIRouter, Header, Path
 
 from app.contracts.proposals import (
     ProposalApprovalActionRequest,
-    ProposalApprovalsEnvelopeResponse,
-    ProposalLineageEnvelopeResponse,
     ProposalStateTransitionEnvelopeResponse,
     ProposalSubmitRequest,
-    ProposalWorkflowEventsEnvelopeResponse,
 )
 from app.middleware.correlation import correlation_id_var
 from app.services.advisory_service_provider import proposal_service
@@ -139,68 +136,5 @@ async def record_client_consent(
         details=request.details,
         related_version_no=request.related_version_no,
         idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/workflow-events",
-    response_model=ProposalWorkflowEventsEnvelopeResponse,
-    summary="Get Proposal Workflow Events",
-    description="Returns the workflow event timeline for a specific advisory proposal.",
-)
-async def get_workflow_events(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalWorkflowEventsEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_workflow_events(
-        proposal_id=proposal_id,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/approvals",
-    response_model=ProposalApprovalsEnvelopeResponse,
-    summary="Get Proposal Approvals",
-    description="Returns approval records already captured for a specific advisory proposal.",
-)
-async def get_approvals(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalApprovalsEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_approvals(
-        proposal_id=proposal_id,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/lineage",
-    response_model=ProposalLineageEnvelopeResponse,
-    summary="Get Proposal Lineage",
-    description="Returns immutable version lineage metadata and hashes for a specific proposal.",
-)
-async def get_proposal_lineage(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalLineageEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_proposal_lineage(
-        proposal_id=proposal_id,
         correlation_id=correlation_id,
     )
