+++ src/app/routers/proposal_workflow_evidence.py
@@ -1,0 +1,74 @@
+from fastapi import APIRouter, Path
+
+from app.contracts.proposals import (
+    ProposalApprovalsEnvelopeResponse,
+    ProposalLineageEnvelopeResponse,
+    ProposalWorkflowEventsEnvelopeResponse,
+)
+from app.middleware.correlation import correlation_id_var
+from app.services.advisory_service_provider import proposal_service
+
+router = APIRouter(prefix="/api/v1/proposals", tags=["proposals"])
+
+
+@router.get(
+    "/{proposal_id}/workflow-events",
+    response_model=ProposalWorkflowEventsEnvelopeResponse,
+    summary="Get Proposal Workflow Events",
+    description="Returns the workflow event timeline for a specific advisory proposal.",
+)
+async def get_workflow_events(
+    proposal_id: str = Path(
+        ...,
+        description="Gateway-visible proposal identifier returned by lotus-advise.",
+        examples=["pp_1"],
+    ),
+) -> ProposalWorkflowEventsEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_workflow_events(
+        proposal_id=proposal_id,
+        correlation_id=correlation_id,
+    )
+
+
+@router.get(
+    "/{proposal_id}/approvals",
+    response_model=ProposalApprovalsEnvelopeResponse,
+    summary="Get Proposal Approvals",
+    description="Returns approval records already captured for a specific advisory proposal.",
+)
+async def get_approvals(
+    proposal_id: str = Path(
+        ...,
+        description="Gateway-visible proposal identifier returned by lotus-advise.",
+        examples=["pp_1"],
+    ),
+) -> ProposalApprovalsEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_approvals(
+        proposal_id=proposal_id,
+        correlation_id=correlation_id,
+    )
+
+
+@router.get(
+    "/{proposal_id}/lineage",
+    response_model=ProposalLineageEnvelopeResponse,
+    summary="Get Proposal Lineage",
+    description="Returns immutable version lineage metadata and hashes for a specific proposal.",
+)
+async def get_proposal_lineage(
+    proposal_id: str = Path(
+        ...,
+        description="Gateway-visible proposal identifier returned by lotus-advise.",
+        examples=["pp_1"],
+    ),
+) -> ProposalLineageEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_proposal_lineage(
+        proposal_id=proposal_id,
+        correlation_id=correlation_id,
+    )
